+++ src/@types/global.d.ts
@@ -1,1 +1,3 @@
 declare const _VERSION_: string;
+declare const _GIT_COMMIT_: string;
+declare const _GIT_MESSAGE_: string;